+++ tests/testthat/test-get_phylopic.R
@@ -13,6 +13,7 @@ test_that("get_phylopic works", {
   expect_error(get_phylopic(uuid = c("1", 2)))
   expect_error(get_phylopic(uuid = NULL))
   expect_error(get_phylopic(uuid = uuid, format = "VHS"))
+  expect_error(get_phylopic(uuid = uuid, preview = "TRUE"))
   
   # 512 was deprecated for format
   expect_warning(get_phylopic(uuid = "c8f71c27-71db-4b34-ac2d-e97fea8762cf",